# docs/swagger.yaml
basePath: /
definitions:
  body_types.GetPayload:
    properties:
      answer_file:
        type: string
      config:
        type: string
      params:
        additionalProperties:
          type: string
        type: object
    type: object
  body_types.GetServiceIP:
    properties:
      namespace:
        type: string
      release_name:
        type: string
    type: object
  body_types.ImportChartRepo:
    properties:
      repo:
        type: string
      repo_name:
        type: string
    type: object
  body_types.InstallUpgradeRequest:
    properties:
      chart:
        type: string
      config:
        additionalProperties: true
        type: object
      namespace:
        type: string
      params:
        additionalProperties:
          type: string
        type: object
      release_name:
        type: string
    type: object
  body_types.NamespaceChartRepo:
    properties:
      namespace:
        type: string
    type: object
  body_types.ServiceDetails:
    properties:
      ip:
        type: string
      port:
        type: integer
      service_name:
        type: string
      service_type:
        type: string
    type: object
  body_types.UninstallChartRepo:
    properties:
      namespace:
        type: string
      release_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    url: https://github.com/Mrpye/helm-api
  description: Helm-api is a CLI application written in Golang that gives the ability
    to perform Install, Uninstall and Upgrade Helm Charts via Rest API endpoint. The
    application can be run as a stand alone application or deployed as a Container.
    Also for convenience there is the  ability to create namespaces and retrieve service
    IPs of the deployed application. GitHub repository at https://github.com/Mrpye/helm-api
  license:
    name: Apache 2.0 licensed
    url: https://github.com/Mrpye/helm-api/blob/main/LICENSE
  title: helm-api
  version: "1.0"
paths:
  /:
    get:
      operationId: check-api-endpoint
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Check API Endpoint
  /add_repo:
    post:
      operationId: add-helm-chart-repo
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.ImportChartRepo'
      produces:
      - application/json
      responses:
        "200":
          description: charts Repo Added
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
      summary: add a helm chart repo
  /create_ns:
    post:
      operationId: create-namespace
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.NamespaceChartRepo'
      produces:
      - application/json
      responses:
        "200":
          description: namespace created
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
      summary: Create Namespace
  /get_config:
    post:
      operationId: get-helm-chart-config
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.GetPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/body_types.InstallUpgradeRequest'
            type: array
        "404":
          description: error
          schema:
            type: string
      summary: get the config for helm chart
  /get_ip:
    post:
      operationId: get-service-ip
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.GetServiceIP'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/body_types.ServiceDetails'
            type: array
        "404":
          description: error
          schema:
            type: string
      summary: Get Service IP
  /install:
    post:
      operationId: install-helm-chart
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.InstallUpgradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: chart installed
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
      summary: Install a helm chart
  /uninstall:
    post:
      operationId: uninstall-helm-chart
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.UninstallChartRepo'
      produces:
      - application/json
      responses:
        "200":
          description: chart uninstalled
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
      summary: uninstall a helm chart
  /upgrade:
    post:
      operationId: upgrade-helm-chart
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/body_types.InstallUpgradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: chart upgraded
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
      summary: Upgrade a helm chart
swagger: "2.0"